# backend/openapi.yaml
openapi: 3.1.0
info:
  title: LangGraph Chatbot API
  description: AI-powered chatbot with conversation history using LangGraph and MongoDB
  version: 1.0.0
paths:
  /auth/generate-otp:
    post:
      tags:
      - Generate OTP
      summary: Send Otp Endpoint
      description: "Generate and send OTP to user's email.\n\nCreates a 6-digit OTP,\
        \ stores it in the database with 5-minute expiration,\nand sends it to the\
        \ user's email address.\n\nGuards: CAPTCHA verification (\xA710), 4-layer\
        \ email validation (\xA74),\nOTP rate limiting (\xA75.1), bounce/complaint\
        \ flag pre-send check (\xA76.3).\n\nArgs:\n    request: FastAPI request (used\
        \ to extract client IP for CAPTCHA)\n    email: User's email address\n   \
        \ captcha_token: Cloudflare Turnstile token (optional in dev mode)\n\nReturns:\n\
        \    Success message confirming OTP was sent"
      operationId: send_otp_endpoint_auth_generate_otp_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_send_otp_endpoint_auth_generate_otp_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/signup/user:
    post:
      tags:
      - User Auth
      summary: Signup User
      description: "Register a new user account.\n\nVerifies the OTP, creates a new\
        \ user with 'student' role, sends welcome email,\nand returns access and refresh\
        \ tokens.\n\nArgs:\n    user_data: User registration data including name,\
        \ email, password, and OTP\n\nReturns:\n    Access token, refresh token, and\
        \ token type\n\nRaises:\n    HTTPException: If OTP is invalid or email already\
        \ registered"
      operationId: signup_user_auth_signup_user_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimplifiedUser'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/signup/admin:
    post:
      tags:
      - Admin Auth
      summary: Signup Admin
      description: "Register a new admin account.\n\nVerifies the OTP, creates a new\
        \ user with 'admin' role, sends welcome email,\nand returns access and refresh\
        \ tokens.\n\nArgs:\n    user_data: User registration data including name,\
        \ email, password, and OTP\n\nReturns:\n    Access token, refresh token, and\
        \ token type\n\nRaises:\n    HTTPException: If OTP is invalid or email already\
        \ registered"
      operationId: signup_admin_auth_signup_admin_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimplifiedUser'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/refresh-token:
    post:
      tags:
      - Token Management
      summary: Refresh Token Endpoint
      description: "Refresh access token using a valid refresh token.\n\nArgs:\n \
        \   refresh_token: The refresh token provided by the user.\n\nReturns:\n \
        \   A new access token and refresh token.\n\nRaises:\n    HTTPException: If\
        \ the refresh token is invalid or expired."
      operationId: refresh_token_endpoint_auth_refresh_token_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
      security:
      - OAuth2PasswordBearer: []
  /auth/login/user:
    post:
      tags:
      - User Auth
      summary: Login User
      description: "Authenticate user and return access tokens.\n\nValidates credentials,\
        \ sends login notification email, and returns JWT tokens.\n\nArgs:\n    form_data:\
        \ OAuth2 form with username (email) and password\n\nReturns:\n    Access token,\
        \ refresh token, and token type\n\nRaises:\n    HTTPException: If credentials\
        \ are invalid"
      operationId: login_user_auth_login_user_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_user_auth_login_user_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login/admin:
    post:
      tags:
      - Admin Auth
      summary: Login Admin
      description: "Authenticate admin and return access tokens.\n\nValidates credentials,\
        \ checks for admin role, sends login notification email, \nand returns JWT\
        \ tokens.\n\nArgs:\n    form_data: OAuth2 form with username (email) and password\n\
        \nReturns:\n    Access token, refresh token, and token type\n\nRaises:\n \
        \   HTTPException: If credentials are invalid or user is not an admin"
      operationId: login_admin_auth_login_admin_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_admin_auth_login_admin_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/password-reset/request:
    post:
      tags:
      - Password Reset
      summary: Password Reset Request
      description: "Request a password reset by sending an OTP to the user's email.\n\
        \nGenerates an OTP and emails it to the user if the account exists.\nAlways\
        \ returns success message for security (prevents email enumeration).\n\nArgs:\n\
        \    email: User's email address\n\nReturns:\n    Success message (always\
        \ returns success for security)"
      operationId: password_reset_request_auth_password_reset_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_password_reset_request_auth_password_reset_request_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/password-reset/verify:
    post:
      tags:
      - Password Reset
      summary: Verify Otp And Reset Password
      description: "Verify OTP and reset the user's password.\n\nValidates the OTP,\
        \ updates the user's password, and sends confirmation email.\n\nArgs:\n  \
        \  data: Email, OTP, and new password\n\nReturns:\n    Success message\n\n\
        Raises:\n    HTTPException: If OTP is invalid or expired"
      operationId: verify_otp_and_reset_password_auth_password_reset_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_login_admin_auth_login_admin_post:
      properties:
        grant_type:
          anyOf:
          - type: string
            pattern: ^password$
          - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          format: password
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          format: password
          title: Client Secret
      type: object
      required:
      - username
      - password
      title: Body_login_admin_auth_login_admin_post
    Body_login_user_auth_login_user_post:
      properties:
        grant_type:
          anyOf:
          - type: string
            pattern: ^password$
          - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          format: password
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          format: password
          title: Client Secret
      type: object
      required:
      - username
      - password
      title: Body_login_user_auth_login_user_post
    Body_password_reset_request_auth_password_reset_request_post:
      properties:
        email:
          type: string
          title: Email
      type: object
      required:
      - email
      title: Body_password_reset_request_auth_password_reset_request_post
    Body_send_otp_endpoint_auth_generate_otp_post:
      properties:
        email:
          type: string
          title: Email
        captcha_token:
          type: string
          title: Captcha Token
          description: Cloudflare Turnstile token
          default: ''
      type: object
      required:
      - email
      title: Body_send_otp_endpoint_auth_generate_otp_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
      description: Generic single-message response body for informational endpoints.
...
    PasswordResetConfirm:
      properties:
        email:
          type: string
          format: email
          title: Email
        otp:
          type: integer
          title: Otp
        new_password:
          type: string
          title: New Password
      type: object
      required:
      - email
      - otp
      - new_password
      title: PasswordResetConfirm
      description: Payload for confirming a password reset with OTP.
...
    SimplifiedUser:
      properties:
        name:
          type: string
          title: Name
          example: John Doe
        email:
          type: string
          format: email
          title: Email
          example: john@example.com
        password:
          type: string
          title: Password
          example: securePass123
        otp:
          type: integer
          title: Otp
          description: 6-digit code
          example: 123456
      type: object
      required:
      - name
      - email
      - password
      - otp
      title: SimplifiedUser
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      - refresh_token
      title: Token
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/login/user